requirements.txt
asgiref==3.5.2
Django==4.1.1
django-ninja==0.19.1
ecdsa==0.18.0
Pillow==9.2.0
pyasn1==0.4.8
pydantic==1.10.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2